User route and controller renamed to users
