+++ api/src/controllers/users.js
@@ -15,7 +15,7 @@ import {
   
     await createAudit({
       apiVersion: 'v1',
-      action: AUDIT_ACTIONS.api.user.getPendingInvitations,
+      action: AUDIT_ACTIONS.api.users.getPendingInvitations,
       data: JSON.stringify({
         message: `user attempting to view pending requests`,
         user: req.auth.user,
@@ -32,7 +32,6 @@ import {
       log.error(requests)
       //  we have the requests made by the user (on other peoples'  behalf), now use github api to see the real status of those requests
       const requestStatuses = await getRequestStatuses(requests)
-      log.info(requestStatuses)
       res.status(200).json(requestStatuses)
   
     } catch (e) {
